tests: A new test prereq for testing chmod -w as root
Some tests depend on not being able to read files after chmod -w. This
doesn't work when running the tests as root.

Change test-lib.sh to test if this works, and if so it sets a new
CHMOD_0000 test prerequisite. The tests that use this previously
failed when run under root.

There was already a test for this in t3600-rm.sh added by Junio C
Hamano in 2283645b85 in 2006. That check now uses the new CHMOD_0000
prerequisite.

There are also two other prerequisites, "POSIXPERM_AND_CHMOD_0000" and
"SYMLINKS_AND_CHMOD_0000". They're being used for tests that failed
and already depended on POSIXPERM or SYMLINKS. They're possibly
redundant, but I don't have access to a system without POSIX
permissions or symlinks so I couldn't test that.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>