pull --rebase: Avoid spurious conflicts and reapplying unnecessary patches
Prior to c85c79279df2c8a583d95449d1029baba41f8660, pull --rebase would run
  git rebase $merge_head
which resulted in a call to
  git format-patch ... --ignore-if-in-upstream $merge_head..$cur_branch

This had two nice qualities when upstream isn't rebased: (1) "only" the
patches in $merge_head..$cur_branch would be applied, and (2) patches
could be dropped/ignored if they had already been applied.  But this did
not work well when upstream is rebased, since in that case
$merge_head..$cur_branch refers to too many commits that would need to be
reapplied, and could result in intentionally dropped commits being
reintroduced.

So the algorithm was changed.  Defining $old_remote_ref to be the most
recent entry in the reflog for @upstream that is an ancestor of
$cur_branch, pull --rebase was changed (over time) to run
  git rebase --onto $merge_head $old_remote_ref
which results in a call to
  git format-patch ... --ignore-if-in-upstream $old_remote_ref..$cur_branch

In the rebased upstream case, this can result in far fewer commits being
included in the rebase, though the fact that $old_remote_ref is an ancestor
of $cur_branch means that format-patch will not know what upstream is and
will not be able to determine if any patches are already upstream.

In the non-rebased upstream case, this new form is usually the same as the
original code.  However, as noted above, the --ignore-if-in-upstream flag
becomes meaningless and all patches will be forced to be reapplied.  Also,
$old_remote_ref can be an ancestor of
   $(git merge-base $merge_head $cur_branch)
meaning that pull --rebase may try to reapply more patches which are
clearly already upstream, without the means to detect that they've already
been applied.  This can be extremely confusing for new users ("git is
giving me lots of conflicts in stuff I didn't even change since my last
push.")

Fix the non-rebased upstream case by ignoring $old_remote_ref whenever it
is contained by $(git merge-base $merge_head $cur_branch).

Signed-off-by: Elijah Newren <newren@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>