imap-send: Code correctness flagged by clang
Clang 1.1 flagged the following issues in imap-send.c, this change
fixes the warnings by moving some code around:

    imap-send.c:548:27: warning: data argument not used by format string [-Wformat-extra-args]
                               cmd->tag, cmd->cmd, cmd->cb.dlen);
                                                   ^

Here the sprintf format didn't use the cmd->cb.dlen argument if
cmd->cb.data was false. Change the code to use a if/else instead of a
two-level ternary to work it. This code was introduced with imap-send
itself in f2561fda.

    imap-send.c:1089:41: warning: conversion specifies type 'unsigned short' but the argument has type 'int' [-Wformat]
                    snprintf(portstr, sizeof(portstr), "%hu", srvc->port);
                                                        ~~^   ~~~~~~~~~~

Here sprintf is being given an int with a %hu format. Cast the
srvc->port to unsigned short to work it. This code was introduced in
94ad2437 to add IPv6 support.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>