t/harness: Work around a TAP::Harness regression
TAP::Harness had a regression in 3.17_01 that'll be fixed in the
upcoming 3.22 release. 3.17_01 - 3.21 have never been released as part
of a non-developer Perl release, but users may have installed it from
the CPAN.

Change t/harness to work around the bug by prefixing tests with ./
before they're passed to TAP::Harness.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>