po/is.po: add Icelandic translation
Translate some non-TEST into Icelandic. This is far from a complete
translation.

Also to use the header order added by msgmerge(1), this'll make
subsequent diffs to it smaller and easier to manage. While I'm at it
add a Language header indicating that the file is in Icelandic.

When we add more *.po files later we should add a Language header to
those as well.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>