gettext tests: test message re-encoding under Perl
Change the Perl Git::I18N tests to test under both a UTF-8 and
ISO-8859-1 locale if available.

Unlike the C and Shell gettext implementations this has no effect on
the output. Locale::Messages doesn't appear to be capable of
re-encoding as the C and Shell implementations are.

Perhaps we need to work around this in the future, e.g. by marking all
gettext messages as UTF-8 with Encoding.pm when they come in, and then
set a binmode() on the STDOUT and STDERR file handles.

But for now I haven't made this into a TODO test.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>