gettext tests: remove old sanity test under NO_GETTEXT
Remove an old sanity test that checked that the share/locale dir
didn't exist if we compiled with NO_GETTEXT=YesPlease. This would
break if the user did:

    git clean -dxf
    make NO_GETTEXT=
    make NO_GETTEXT=YesPlease
    make test

This test was something to test very basic sanity back when I was
first developing the gettext series, but it's not needed anymore. We
don't have similar tests for other parts of git to check that we
didn't build something with a given feature disabled.

Reported-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>