gettextize: git-status basic messages
Ripped out as-is from the "Add initial C, Shell and Perl gettext
translations" patch that was part of the gettext RFC for a long time
and has already been reviewed by a lot of eyeballs.

The only change is the removal of two messages that didn't really
belong: _("#\t") and _("%-20s%s\n").

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>