gettext: Add po/README file documenting Git's gettext
Add a po/README file. This documentation is targated at translators
and maintainers of git.git.

Currently it describes how to add new PO files and how to update
them. But in the future we want to add more things to it, e.g. how and
what to add in TRANSLATOR comments, how to make the source more
gettext friendly etc.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>