po/hi.po: add Hindi Translation
Translate a few Git messages to Standard Hindi. It uses a phonetic
script called Devanagari. Focus especially on transforming grammar and
tense in a sane manner to serve as a guideline for future translators.

Note that Devanagari rendering is broken on many applications by
default at this time- the latest Emacs, Vim and Chrome render it
incorrectly.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Acked-by: Antriksh Pany <antriksh.pany@gmail.com>
Acked-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>