Makefile: move "Platform specific tweaks" above LIB_{H,OBJS}
Change the Makefile so that the "Platform specific tweaks" section
comes before the assignments to LIB_H and LIB_OBJS.

In the ab/i18n series I only want to build gettext.o (by adding it to
LIB_OBJS) if NO_GETTEXT is unset. It's not possible to do that without
an ugly hack if we haven't applied our platform specific tweaks before
LIB_{H,OBJS} gets assigned to.

See <201008140002.40587.j6t@kdbg.org> (subject: "[PATCH] Do not build
i18n on Windows.") for Johannes's original report, and my follow-up in
<AANLkTiku5R+idX-C8f0AcCikBLmfEb5ZEhdft+CSRzU0@mail.gmail.com> where I
suggested that the problem be solved in the manner of this patch.

Reported-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>