vcs-svn: prepare to eliminate repo_tree structure
Currently svn-fe processes each commit in two stages: first decide on
the correct content for all paths and export the relevant blobs, then
export a commit with the result.

But we can keep less state and simplify svn-fe a great deal by
exporting the commit in one step: use 'inline' blobs for each path and
remember nothing.  This way, the repo_tree structure could be
eliminated, and we would get support for incremental imports 'for
free'.

Reorganize handle_node() along these lines.  This is just a code
cleanup; the changes in repo_tree will come in later patches.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>