cache.h: include gettext.h
Change cache.h to include gettext.h. In previous incarnations of the
gettext series I included gettext.h in builtin.h, then changed
builtin/*.c files to include that when they didn't already.

It was pointed out by Junio and others that it makes more sense to
have the inclusion here, since gettext is going to be used everywhere.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>