po/is.po: initial Icelandic translation
Add a partially completed Icelandic translation, around 60% of strings
have been translated. This is far from perfect but gives us some
initial set to work with, the main init/status/commit commands were
the main focus of this translation.

This translation uses the terminology for the core commands described
in po/is.README.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>